+++ languages/en/tl_news_category.php
@@ -21,6 +21,7 @@ $GLOBALS['TL_LANG']['tl_news_category']['alias']             = array('Category a
 $GLOBALS['TL_LANG']['tl_news_category']['cssClass']          = array('CSS class', 'Here you can enter a CSS class that will be added to the category in front end.');
 $GLOBALS['TL_LANG']['tl_news_category']['hideInList']        = array('Hide in list/archive module', 'Do not display category in the news list/archive module (affects only the <em>news_</em> templates).');
 $GLOBALS['TL_LANG']['tl_news_category']['hideInRelatedList'] = array('Hide in related list module', 'Do not display category in the related news list module (affects only the <em>news_</em> templates).');
+$GLOBALS['TL_LANG']['tl_news_category']['hideInReader']      = array('Hide in reader module', 'Do not display category in the news reader module (affects only the <em>news_</em> templates).');
 $GLOBALS['TL_LANG']['tl_news_category']['jumpTo']            = array('Redirect page', 'Here you can choose the page to which visitors will be redirected when clicking a category link in the news template.');
 $GLOBALS['TL_LANG']['tl_news_category']['published']         = array('Publish category', 'Make the news category publicly visible on the website.');