+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 2.8.0 stable (????-??-??)
 ### Improved
 - Allow to hide category in the news list module
 - Allow to hide category in the related news list module
+- Allow to hide category in the news reader module
 
 
 Version 2.7.5 stable (2016-10-20)